Start revamp of test suit
